change noise to float

+++ layers/gaussian_noise.py
@@ -27,6 +27,6 @@ class GaussianNoise(nn.Module):
     def forward(self, x):
         if self.training and self.sigma != 0:
             scale = self.sigma * x.detach() if self.is_relative_detach else self.sigma * x
-            sampled_noise = self.noise.repeat(*x.size()).normal_() * scale
+            sampled_noise = self.noise.repeat(*x.size()).float().normal_() * scale
             x = x + sampled_noise
         return x
